herodium/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
psutil==7.2.1
pyClamd==0.4.0
PyYAML==6.0.3
requests==2.32.5
sdnotify==0.3.2
urllib3==2.6.3
watchdog==6.0.0
